# internal/apiserver/openapi.yaml
openapi: 3.0.3
info:
  title: "gNMIc Operator REST API"
  version: "0.0.2"
paths:
  /clusters/:namespace/:name/plan:
    get:
      summary: "Get cluster plan."
      operationId: "getClusterPlan"
      responses:
        '200':
          description: "ClusterPlan returned"
        '404':
          description: "No plan is cached for this Cluster"
  /api/v1/namespaces/:namespace/targetsources/:name/refresh:
    post:
      summary: "Request an immediate discovery run for a TargetSource."
      description: >
        Notify-only. The body carries no target data; it is only an input to
        signature verification. A successful call annotates the TargetSource
        and the controller performs an ordinary full run. Calls within the
        TargetSource's webhook.debounce window of the previous one are
        acknowledged without scheduling another run.
      operationId: "refreshTargetSource"
      security:
        - bearerAuth: []
        - signature: []
      requestBody:
        required: false
        description: Opaque. Forwarded to HMAC verification when signature auth is configured.
        content:
          '*/*':
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: "A run was requested"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '200':
          description: "Debounced: a run was requested less than webhook.debounce ago"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: "Authentication failed"
        '404':
          description: "The TargetSource does not exist or has no webhook enabled"

components:
  schemas:
    RefreshResponse:
      type: object
      required:
        - requestedAt
        - debounced
      properties:
        requestedAt:
          type: string
          format: date-time
          description: When a run was last requested through this endpoint.
        debounced:
          type: boolean
          description: True when this call did not schedule a new run because one was requested within the debounce window.

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Compared in constant time against the Secret named by spec.webhook.auth.bearer.
    signature:
      name: X-Hook-Signature
      type: apiKey
      in: header
      description: Hex HMAC of the request body, optionally prefixed with "sha256=" or "sha512=", keyed by the Secret named by spec.webhook.auth.signature.